Moved grammar to it's own class. Consequently broke some tests.
