+++ CMakeLists.txt
@@ -40,6 +40,7 @@ include_directories(thirdparty/simdjson)
 include_directories(src)
 include_directories(src/Grids)
 include_directories(src/Utlities)
+include_directories(src/Grammar)
 include_directories(src/CellComplexes)
 include_directories(src/CellComplexes/Utils)
 include_directories(src/Solvers)
